Moved tests into proper home
Tests rightly belong in tree tests not xml-parser tests.
